requirements.txt
duckdb==0.7.0
streamlit==1.18.0
geopy==2.3.0
minio==7.1.13
pandas==1.5.3
airflow-provider-duckdb==0.0.2
astro-sdk-python==1.5.0
altair==4.0.0
apache-airflow-providers-postgres
pandas

